backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
python-dotenv==1.0.1
SQLAlchemy==2.0.40 